+++ lib/mailer.dart
@@ -8,7 +8,8 @@ import 'src/smtp/smtp_client.dart';
 export 'src/entities.dart';
 export 'legacy.dart';
 
-Future<List<SendReport>> send(Message message, SmtpServer smtpServer) {
+Future<List<SendReport>> send(Message message, SmtpServer smtpServer,
+                             {Duration timeout}) {
   var client = new SmtpClient(smtpServer);
-  return client.send(message);
+  return client.send(message, timeout: timeout);
 }